% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pc_scores.R
\name{get_pc_scores}
\alias{get_pc_scores}
\title{Calculate the PC scores}
\usage{
get_pc_scores(fastpca_out)
}
\arguments{
\item{fastpca_out}{List: output from running FastPCA::FastPCA with elements U, S, and Vh}
}
\value{
Matrix: contains the scores associated with each of the PCs
}
\description{
Calculate the PC scores
}
