% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCs.R
\name{plotPCs}
\alias{plotPCs}
\title{Function for plotting FastPCA objects}
\usage{
plotPCs(PCA_ob, PC1 = 1, PC2 = 2, label = TRUE)
}
\arguments{
\item{PCA_ob}{Output from FastPCA. Required}

\item{PC1}{Principal component that will be plotted on x-axis. Default is 1}

\item{PC2}{Principal component that will be plotted on y-axis. Default is 2}

\item{label}{Parameter determining whether points should be labeled. Labels
are determined by the row names of the original object which are determined
by column names of the original data. Default is TRUE}
}
\value{
Plot of the two specified principal components
}
\description{
Function for plotting FastPCA objects
}
\details{
Simple plotting function for FastPCA output. Takes the FastPCA
object and PCs of interest as input. Also supports labeling.
}
\examples{
#Create a simulated matrix
x = matrix(runif(900,0,20), nrow = 30, ncol = 30)
#Get the first 10 PCs
PCs <- FastPCA(x, PCs = 10)
}
