% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_matrix.R
\name{prep_matrix}
\alias{prep_matrix}
\title{Perform transpose and scaling of matrix}
\usage{
prep_matrix(mat, log2 = TRUE, transpose = FALSE, scale = TRUE)
}
\arguments{
\item{mat}{A numeric R matrix.}

\item{transpose}{Boolean. Whether the matrix needs to be transposed. If starting with samples as columns, set to \code{TRUE}. Default is \code{FALSE}}

\item{scale}{Boolean. Whether to center and scale the matrix. Default to \code{TRUE}}
}
\value{
A matrix that has been rotated (transposed) and scaled if needed
}
\description{
This function uses PyTorch implementation of Randomized SVD to compute
the singular value decomposition.
}
\examples{
\dontrun{
  #need to have environment first
  reticulate::use_condaenv("FastPCA")

  # Create a sample R matrix (e.g., 20 samples, 100 features)
  # Ensure values are positive for log transform.
  set.seed(123)
  test_data <- matrix(runif(20 * 100, min=1, max=100), nrow = 20, ncol = 100)
  colnames(test_data) <- paste0("Feature", 1:ncol(test_data))
  rownames(test_data) <- paste0("Sample", 1:nrow(test_data))

  print(paste("Original R matrix dimensions:", paste(dim(test_data), collapse = "x")))

  # Perform Randomized SVD
  prepped_matrix <- FastPCA::prep_matrix(test_data, log2 = TRUE,
                                  transpose = FALSE, scale = TRUE)
}
}
