% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtorch_randomized_svd.r
\name{rtorch_randomized_svd}
\alias{rtorch_randomized_svd}
\title{Randomized Singular Vector Decomposition using torch}
\usage{
rtorch_randomized_svd(A_mat, k, p, q_iter, device = "CPU", cores = 1)
}
\arguments{
\item{A_mat}{matrix passed from R}

\item{k}{Integer. Number of dimensions to calculate the singular vectors on}

\item{p}{Intger. Oversampling dimensions}

\item{q_iter}{Integer. Number of power iterations to perform}

\item{device}{Character. Either "CPU" or "GPU depending on what user selects and availability}

\item{cores}{Integer. Number of CPU cores to use for matrix operations}
}
\value{
a list with U, S, and Vh
}
\description{
Randomized Singular Vector Decomposition using torch
}
