% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_FastPCA_env.R
\name{start_FastPCA_env}
\alias{start_FastPCA_env}
\title{Activate python environment for FastPCA}
\usage{
start_FastPCA_env(
  method = c("conda", "virtualenv"),
  envname = "FastPCA",
  required = FALSE
)
}
\arguments{
\item{method}{Character string. The method to use for environment creation.
Can be "conda" (default) or "virtualenv".}

\item{envname}{Character string. The name of the Python environment to create/use.
Defaults to "FastPCA".}

\item{required}{boolean passed to \code{reticulate}}
}
\value{
Invisibly returns \code{TRUE} if setup is successful, \code{FALSE} otherwise.
}
\description{
Activate python environment for FastPCA
}
\examples{
\dontrun{
  #for conda
  setup_py_env(method = "conda", envname = "FastPCA",
    python_version = "3.9", backend = "all", cuda = FALSE)
  start_fpca_env
}
}
