% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{umap}
\alias{umap}
\title{umap}
\usage{
umap(
  pc_scores,
  n_neighbors = 15,
  n_components = 2,
  metric = "euclidean",
  min_dist = 0.1,
  cores = 4,
  seed = -1,
  densemap = TRUE,
  dense_lambda = 2,
  verbose = FALSE
)
}
\arguments{
\item{pc_scores}{Matrix: exported from \code{get_pc_scores()}}

\item{n_neighbors}{Integer: number of nearest neighbors to use}

\item{n_components}{Integer: number of UMAP dimensions to calculate}

\item{metric}{Character: How to calculate similarity. See details below for more informaation.}

\item{min_dist}{Numeric: minimum distance to consider}

\item{cores}{Integer: Number of cores to use for UMAP calculation}

\item{seed}{Integer: reproducibility seed. If cores > 1 or seed equals \code{-1}, seed will be ignored.}

\item{densemap}{Boolean: whether to use DensMAP for local densities}

\item{dense_lambda}{Numeric: value to apply for local density. Default: 2. higher values prioritize local density while low values are closer to typical UMAP}

\item{verbose}{Boolean: whether to be verbose in function calls}
}
\value{
matrix with UMAP reductions
}
\description{
umap
}
\details{
\code{metric}
There are many metrics that are supported in the python implementation. Here
are the list in the \href{https://umap-learn.readthedocs.io/en/latest/parameters.html#metric}{documentation}
for the umap function;
\itemize{
\item Minkowski syle metrics: "euclidean", "manhattan", "chebyshev", "minkowski";
\item Miscellaneous spatial metrics: "canberra", "braycurtis", "haversine";
\item Normalized spatial metrics: "mahalanobis", "wminkowski", "seuclidean";
\item Angular and correlation metrics: "cosine", "correlation";
\item Metrics for binary data: "hamming", "jaccard", "dice", "russellrao", "kulsinski", "rogerstanimoto", "sokalmichener", "sokalsneath", "yule"
}
}
